- AssetManager pch 사용 중지 - MeshPartsData에 대한 Visitor를 별도 설계하여 모델파일 로딩 기능 재구현 - Blend Index, Weight 로딩 관련 버그 수정

+++ AssetManager/MathematicalHelper.cpp
@@ -1,11 +1,11 @@
-#include "pch.h"
+
 #include "MathematicalHelper.h"
 
 using namespace DirectX;
 
 XMFLOAT3 Normalize(const XMFLOAT3& Vector)
 {
-	float Length = sqrt(Vector.x * Vector.x + Vector.y * Vector.y + Vector.z * Vector.z);
+	float Length = sqrtf(Vector.x * Vector.x + Vector.y * Vector.y + Vector.z * Vector.z);
 	return XMFLOAT3{ Vector.x / Length, Vector.y / Length, Vector.z / Length };
 }
